refactor profile layout and water info calculations for improved readability and precision

+++ app/src/main/java/com/duzceders/aicaltracker/product/models/User.java
@@ -26,8 +26,8 @@ public class User {
     private int bmr;
     private int daily_calorie_needs;
     private int daily_calorie_needs_left;
-    private int daily_water_needs_liters;
-    private int daily_water_needs_left_liters;
+    private double daily_water_needs_liters;
+    private double daily_water_needs_left_liters;
 
     private CalorieGoals calorie_goals;
     private DailyMacros daily_macros;
